corrected PolygonTMO's rotation
fck u float

+++ graphics-editor-cgs/HorizontalLine.cs
@@ -3,11 +3,11 @@ namespace graphics_editor_cgs
 {
     public class HorizontalLine
     {
-        public float xl;
-        public float xr;
-        public float y;
+        public int xl;
+        public int xr;
+        public int y;
 
-        public HorizontalLine(float xl, float xr, float y) : this()
+        public HorizontalLine(int xl, int xr, int y) : this()
         {
             this.xl = xl;
             this.xr = xr;
